requirements.txt
numpy
pandas
scipy
spacy
nltk
scikit-learn
transformers
datasets
evaluate
matplotlib
seaborn
plotly
tqdm
python-dotenv
black
torch
pytorch-lightning
pyyaml
dvc
checklist==0.0.11
rouge_score
bert_score
num2words
unbabel-comet